WIP best performance...
using a fixed number of threads instead of hardware_concurrency is for
sanity's sake -- sometimes android returns a different answer, maybe
related to thermal issues?

Fullscreen disabled for the moment to force a smaller image in testing
(1280x960 instead of 1620x1080).

Using BGR in the OpencvCameraView, because -- while it will make the
preview image look inverted -- I'd rather save the cycles.