Downgrade sdk/ndk versions to fix fdroid build.
This is not a case of "fdroid doesn't work" -- it's a build system
problem, and the fdroid builds are catching it. A gradle upgrade might
resolve it, but also might not. For now, I'll go the somewhat tortured
route of building against 33+25.2 when we need new builds in the play
store, and leaving the checked in build.gradle like this.

:grimacing: