5000 is too large for inline-threshold
The default is way too low, so 5000 was an improvement over that. But
1500 ~= 2000 > 1000 > 5000 > default (225?) seems to be the case in my
tests (more to learn here at some point...)

The main relevant code in question is the symbol/color decoder loop
(after scan and extract).

-o2 vs -o3 (RelWithDebInfo vs Release) confounds things a little bit,
but less than I expected. We'll let the build system decide that one for
now (but make sure we pass the options regardless)

-flto=thin seems nice to have regardless.